// src/Utils/Interfaces/IDocker.ts
// Generated by https://quicktype.io

export interface DockerImage {
    Id: string;
    Names: string[];
    Image: string;
    ImageID: string;
    Command: string;
    Created: number;
    Ports: Port[];
    Labels: Labels;
    State: string;
    Status: string;
    HostConfig: HostConfig;
    NetworkSettings: NetworkSettings;
    Mounts: Mount[];
}

export interface HostConfig {
    NetworkMode: string;
}

export interface Labels {}

export interface Mount {
    Type: string;
    Name?: string;
    Source: string;
    Destination: string;
    Driver?: string;
    Mode: string;
    RW: boolean;
    Propagation: string;
}

export interface NetworkSettings {
    Networks: Networks;
}

export interface Networks {
    bridge: Bridge;
}

export interface Bridge {
    IPAMConfig: null;
    Links: null;
    Aliases: null;
    NetworkID: string;
    EndpointID: string;
    Gateway: string;
    IPAddress: string;
    IPPrefixLen: number;
    IPv6Gateway: string;
    GlobalIPv6Address: string;
    GlobalIPv6PrefixLen: number;
    MacAddress: string;
    DriverOpts: null;
}

export interface Port {
    IP?: string;
    PrivatePort: number;
    PublicPort?: number;
    Type: string;
}